+++ src/app/api/login/route.ts
@@ -4,7 +4,7 @@ import { cookies } from 'next/headers';
 import { routeResponse } from '@/app/api/_interceptor/routeResponse';
 import { GatewayError } from '@/app/api/_interceptor/error/classes';
 import { COOKIE } from '@/app/api/_interceptor/utils/cookieUtils';
-import getRfTokenFromSetCookie from '@/utils/getRfTokenFromSetCookie';
+import getRfTokenFromSetCookie from '@/utils/auth/getRfTokenFromSetCookie';
 
 export async function POST(req: NextRequest) {
   const loginRequest = await req.json();
